Extract reports extension
