+++ lib/PhpBench.php
@@ -18,6 +18,7 @@ use PhpBench\DependencyInjection\Container;
 use PhpBench\Exception\ConfigurationPreProcessingError;
 use PhpBench\Extension\CoreExtension;
 use PhpBench\Extension\ExpressionExtension;
+use PhpBench\Extension\ReportExtension;
 use PhpBench\Extension\RunnerExtension;
 use PhpBench\Extensions\XDebug\XDebugExtension;
 use PhpBench\Json\JsonDecoder;
@@ -60,10 +61,12 @@ class PhpBench
             $autoloader->register(true);
         }
 
-        $extensions = $config['extensions'];
-        $extensions[] = CoreExtension::class;
-        $extensions[] = RunnerExtension::class;
-        $extensions[] = ExpressionExtension::class;
+        $extensions = array_merge([
+            CoreExtension::class,
+            RunnerExtension::class,
+            ReportExtension::class,
+            ExpressionExtension::class,
+        ], $config['extensions']);
 
         if (extension_loaded('xdebug')) {
             $extensions[] = XDebugExtension::class;
