+++ tests/IntegrationTestCase.php
@@ -5,6 +5,7 @@ namespace PhpBench\Tests;
 use PhpBench\DependencyInjection\Container;
 use PhpBench\Extension\CoreExtension;
 use PhpBench\Extension\ExpressionExtension;
+use PhpBench\Extension\ReportExtension;
 use PhpBench\Extension\RunnerExtension;
 use PhpBench\Tests\Util\Workspace;
 
@@ -24,6 +25,7 @@ class IntegrationTestCase extends TestCase
         })(new Container([
             ExpressionExtension::class,
             CoreExtension::class,
+            ReportExtension::class,
             RunnerExtension::class,
         ], array_merge([
             ExpressionExtension::PARAM_SYNTAX_HIGHLIGHTING => false,
